Move logging to its own crate
Nicely isolates the global
